requirements.txt
fastapi
uvicorn
jinja2
python-multipart
pydantic
langchain
langchain-community
langchain-ollama
langchain-huggingface
faiss-cpu
sentence-transformers
python-dotenv
pypdf